rename project_id column of Task model to supertask_id
the parent task of a task is a supertask, which is sometimes refered to as a project. but supertask_id is faster for brain!

remember that a project can be a task again, if it has a parent task, in which case, supertask_id is not null.
project and task both share the same model. it is just a naming convenience;

the relationship between task and project is like the relation betweem employee and employer, in which both sides are instances of Person model. Here, both are instanecs of Task model.